requirements.txt
# Complete Supreme Oracle System Requirements

# Deep Learning & AI
torch>=2.0.0
transformers>=4.21.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Data Processing
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0

# Web Dashboard & API
Flask>=2.3.0
Flask-SocketIO>=5.3.0
python-socketio>=5.8.0
eventlet>=0.33.0

# AWS Bedrock AI
boto3>=1.26.0
botocore>=1.29.0

# API & Network
requests>=2.26.0
python-dotenv>=0.19.0
cloudscraper>=1.2.71
websocket-client>=1.4.0
urllib3>=1.26.0

# Advanced Analysis & Security
ripser>=0.6.0  # For TDA analysis
cryptography>=3.4.0
hashlib2>=1.0.1

# System utilities
psutil>=5.8.0
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0

